feat/AllowCamelCaseCustomEvent: add logic in jsx-dom.ts (#82)
HOW:
We consider an event to be a browser's own event if, when converted to lowercase, it belongs to the Window Event Attributes set.

E.g. onclick, onClick, onmouseover, onMouseOver, etc. Window Event Attributes => https://www.w3schools.com/tags/ref_eventattributes.asp

Otherwise we consider it a custom event.  E.g. onMyCustomEvent

So let's trim the prefix "on" and lowercase the first character and add the listener "myCustomEvent".
except for the first character, we keep the event name case

Co-authored-by: Delgado Trueba <delgadotrueba@outlook.com>